+++ iextrading4j-api/src/main/java/pl/zankowski/iextrading4j/api/refdata/v1/SymbolMapping.java
@@ -8,7 +8,7 @@ import com.google.common.base.Objects;
 
 import java.io.Serializable;
 
-@JsonPropertyOrder({"symbol", "exchange", "region"})
+@JsonPropertyOrder({"symbol", "exchange", "region", "iexId"})
 public class SymbolMapping implements Serializable {
 
     private static final long serialVersionUID = 7796962782222356844L;
@@ -16,15 +16,18 @@ public class SymbolMapping implements Serializable {
     private final String symbol;
     private final String exchange;
     private final String region;
+    private final String iexId;
 
     @JsonCreator
     public SymbolMapping(
             @JsonProperty("symbol") final String symbol,
             @JsonProperty("exchange") final String exchange,
-            @JsonProperty("region") final String region) {
+            @JsonProperty("region") final String region,
+            @JsonProperty("iexId") final String iexId) {
         this.symbol = symbol;
         this.exchange = exchange;
         this.region = region;
+        this.iexId = iexId;
     }
 
     public String getSymbol() {
@@ -39,6 +42,10 @@ public class SymbolMapping implements Serializable {
         return region;
     }
 
+    public String getIexId() {
+        return iexId;
+    }
+
     @Override
     public boolean equals(final Object o) {
         if (this == o) {
@@ -50,12 +57,13 @@ public class SymbolMapping implements Serializable {
         final SymbolMapping that = (SymbolMapping) o;
         return Objects.equal(symbol, that.symbol) &&
                 Objects.equal(exchange, that.exchange) &&
-                Objects.equal(region, that.region);
+                Objects.equal(region, that.region) &&
+                Objects.equal(iexId, that.iexId);
     }
 
     @Override
     public int hashCode() {
-        return Objects.hashCode(symbol, exchange, region);
+        return Objects.hashCode(symbol, exchange, region, iexId);
     }
 
     @Override
@@ -64,6 +72,7 @@ public class SymbolMapping implements Serializable {
                 .add("symbol", symbol)
                 .add("exchange", exchange)
                 .add("region", region)
+                .add("iexId", iexId)
                 .toString();
     }