+++ iextrading4j-client/src/main/java/pl/zankowski/iextrading4j/client/rest/request/stocks/ListRequestBuilder.java
@@ -3,14 +3,12 @@ package pl.zankowski.iextrading4j.client.rest.request.stocks;
 import pl.zankowski.iextrading4j.api.stocks.Quote;
 import pl.zankowski.iextrading4j.client.rest.manager.RestRequest;
 import pl.zankowski.iextrading4j.client.rest.manager.RestRequestBuilder;
-import pl.zankowski.iextrading4j.client.rest.request.IEXApiRestRequest;
 import pl.zankowski.iextrading4j.client.rest.request.IEXCloudV1RestRequest;
-import pl.zankowski.iextrading4j.client.rest.request.IRestRequestBuilder;
 
 import javax.ws.rs.core.GenericType;
 import java.util.List;
 
-public class ListRequestBuilder implements IEXApiRestRequest<List<Quote>>, IEXCloudV1RestRequest<List<Quote>> {
+public class ListRequestBuilder implements IEXCloudV1RestRequest<List<Quote>> {
 
     private ListType listType;