+++ iextrading4j-client/src/main/java/pl/zankowski/iextrading4j/client/rest/request/stocks/TodayIposRequestBuilder.java
@@ -3,11 +3,9 @@ package pl.zankowski.iextrading4j.client.rest.request.stocks;
 import pl.zankowski.iextrading4j.api.stocks.TodayIpos;
 import pl.zankowski.iextrading4j.client.rest.manager.RestRequest;
 import pl.zankowski.iextrading4j.client.rest.manager.RestRequestBuilder;
-import pl.zankowski.iextrading4j.client.rest.request.IEXApiRestRequest;
 import pl.zankowski.iextrading4j.client.rest.request.IEXCloudV1RestRequest;
-import pl.zankowski.iextrading4j.client.rest.request.IRestRequestBuilder;
 
-public class TodayIposRequestBuilder implements IEXApiRestRequest<TodayIpos>, IEXCloudV1RestRequest<TodayIpos> {
+public class TodayIposRequestBuilder implements IEXCloudV1RestRequest<TodayIpos> {
 
     @Override
     public RestRequest<TodayIpos> build() {