+++ iextrading4j-test/src/test/java/pl/zankowski/iextrading4j/test/rest/v1/refdata/IsinServiceTest.java
@@ -37,6 +37,7 @@ public class IsinServiceTest extends BaseIEXCloudV1ServiceTest {
         assertThat(symbolMapping.getSymbol()).isEqualTo("foo");
         assertThat(symbolMapping.getExchange()).isEqualTo("NAS");
         assertThat(symbolMapping.getRegion()).isEqualTo("US");
+        assertThat(symbolMapping.getIexId()).isEqualTo("IEX_4D48333344362D52");
     }
 
 }